chore: move chromium to src/server/chromium, enfore installer deps
