+++ src/client/cdpSession.ts
@@ -16,7 +16,7 @@
 
 import { CDPSessionChannel, CDPSessionInitializer } from '../protocol/channels';
 import { ChannelOwner } from './channelOwner';
-import { Protocol } from '../chromium/protocol';
+import { Protocol } from '../server/chromium/protocol';
 
 export class CDPSession extends ChannelOwner<CDPSessionChannel, CDPSessionInitializer> {
   static from(cdpSession: CDPSessionChannel): CDPSession {
