+++ src/install/browserPaths.ts
@@ -18,7 +18,7 @@
 import { execSync } from 'child_process';
 import * as os from 'os';
 import * as path from 'path';
-import { getUbuntuVersionSync } from '../helper';
+import { getUbuntuVersionSync } from '../utils/ubuntuVersion';
 import { getFromENV } from '../utils/utils';
 
 export type BrowserName = 'chromium'|'webkit'|'firefox';
