+++ src/server/chromium/crConnection.ts
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-import { assert } from '../utils/utils';
-import { ConnectionTransport, ProtocolRequest, ProtocolResponse } from '../transport';
+import { assert } from '../../utils/utils';
+import { ConnectionTransport, ProtocolRequest, ProtocolResponse } from '../../transport';
 import { Protocol } from './protocol';
 import { EventEmitter } from 'events';
-import { rewriteErrorMessage } from '../utils/stackTrace';
-import { debugLogger } from '../utils/debugLogger';
+import { rewriteErrorMessage } from '../../utils/stackTrace';
+import { debugLogger } from '../../utils/debugLogger';
 
 export const ConnectionEvents = {
   Disconnected: Symbol('ConnectionEvents.Disconnected')
