+++ src/server/chromium/crExecutionContext.ts
@@ -18,10 +18,10 @@
 import { CRSession } from './crConnection';
 import { getExceptionMessage, releaseObject } from './crProtocolHelper';
 import { Protocol } from './protocol';
-import * as js from '../javascript';
-import * as sourceMap from '../utils/sourceMap';
-import { rewriteErrorMessage } from '../utils/stackTrace';
-import { parseEvaluationResultValue } from '../common/utilityScriptSerializers';
+import * as js from '../../javascript';
+import * as sourceMap from '../../utils/sourceMap';
+import { rewriteErrorMessage } from '../../utils/stackTrace';
+import { parseEvaluationResultValue } from '../../common/utilityScriptSerializers';
 
 export class CRExecutionContext implements js.ExecutionContextDelegate {
   _client: CRSession;
