+++ utils/protocol-types-generator/index.js
@@ -15,15 +15,11 @@ async function generateProtocol(name, executablePath) {
 }
 
 async function generateChromiumProtocol(executablePath) {
-  const outputPath = path.join(__dirname, '..', '..', 'src', 'chromium', 'protocol.ts');
+  const outputPath = path.join(__dirname, '..', '..', 'src', 'server', 'chromium', 'protocol.ts');
+  process.env.PLAYWRIGHT_CHROMIUM_DEBUG_PORT = '9339';
   const playwright = require('../../index').chromium;
-  const defaultArgs = playwright._defaultArgs.bind(playwright);
-  playwright._defaultArgs = (...args) => {
-    const result = defaultArgs(...args);
-    result.push('--remote-debugging-port=9339');
-    return result;
-  };
   const browser = await playwright.launch({ executablePath });
+  delete process.env.PLAYWRIGHT_CHROMIUM_DEBUG_PORT;
   const page = await browser.newPage();
   await page.goto(`http://localhost:9339/json/protocol`);
   const json = JSON.parse(await page.evaluate(() => document.documentElement.innerText));
